pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "financial-statement-fraud-detection"
version = "0.1.0"
description = "End-to-end machine learning project for detecting financial statement fraud from Vietnamese financial statement data."
readme = "README.md"
requires-python = ">=3.10,<3.13"
license = { text = "MIT" }
authors = [
  { name = "Your Name" }
]
keywords = [
  "machine-learning",
  "fraud-detection",
  "financial-statements",
  "xgboost",
  "lightgbm",
  "fastapi",
  "mlops"
]
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Intended Audience :: Science/Research",
  "Intended Audience :: Developers",
  "Topic :: Scientific/Engineering :: Artificial Intelligence"
]

dependencies = [
  "numpy>=1.24,<3.0",
  "pandas>=2.0,<3.0",
  "scikit-learn>=1.3,<2.0",
  "scipy>=1.10,<2.0",

  "matplotlib>=3.7,<4.0",
  "seaborn>=0.12,<1.0",

  "openpyxl>=3.1,<4.0",
  "xlsxwriter>=3.1,<4.0",

  "vnstock>=3.0.0",
  "xgboost>=2.0,<3.0",
  "lightgbm>=4.0,<5.0",

  "fastapi>=0.110,<1.0",
  "uvicorn[standard]>=0.29,<1.0",
  "pydantic>=2.6,<3.0",
  "python-dotenv>=1.0,<2.0",

  "mlflow>=2.12,<3.0",
  "joblib>=1.3,<2.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0,<9.0",
  "pytest-cov>=4.1,<6.0",
  "httpx>=0.27,<1.0",
  "black>=24.0,<25.0",
  "ruff>=0.4,<1.0",
  "mypy>=1.8,<2.0",
  "pre-commit>=3.7,<4.0"
]

notebooks = [
  "jupyter>=1.0,<2.0",
  "ipykernel>=6.0,<7.0",
  "notebook>=7.0,<8.0"
]

serving = [
  "gunicorn>=21.2,<23.0"
]

all = [
  "financial-statement-fraud-detection[dev,notebooks,serving]"
]

[project.urls]
Homepage = "https://github.com/your-username/Financial-Statement-Fraud-Detection"
Repository = "https://github.com/your-username/Financial-Statement-Fraud-Detection"
Issues = "https://github.com/your-username/Financial-Statement-Fraud-Detection/issues"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "api*"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
addopts = "-q --disable-warnings"
filterwarnings = [
  "ignore::UserWarning",
  "ignore::FutureWarning"
]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true